Ajsutes em Movie , auth e Vote

+++ Application/DTOs/Movie/MovieDetailsResponse.cs
@@ -9,7 +9,7 @@
         public required string Genres { get; set; }
         public float Duration { get; set; }
         public float AverageScore { get; set; }
-        public float TotalVotes { get; set; }
+        public int TotalVotes { get; set; }
         public required List<string> Cast { get; set; }
         public required List<string> Directors { get; set; }
         public int ReleasedYear { get; set; }
